+++ Config/encription_config.php
@@ -1,0 +1,7 @@
+<?php
+
+return array(
+    'method' => 'AES-256-CBC',
+    'key' => '37360191345215',
+    'identifier' => 'pukoframework',
+);
